shld overflow fixed, the bug was pressent only on debug mode

+++ src/emu.rs
@@ -1665,7 +1665,7 @@ impl Emu {
         let mut storage0:u64 = value0;
         let mut counter:u64 = pcounter;
 
-        self.flags.f_cf = get_bit!(value0, (size as u64) - counter) == 1;
+        self.flags.f_cf = get_bit!(value0, size - counter as u8) == 1;
 
         if pcounter >= size as u64 {
             counter = pcounter - size as u64;
